Exam mode: Move the exam navigation bar next to its only consumer
ExamNavigationBarComponent lived under exam/overview/, which is the student
exam conduction area, but it is instructor-only: its single renderer is
manage/student-exams/student-exam-timeline/student-exam-timeline.component.html,
which passes examTimeLineView="true". Nothing student-facing uses it — the
student conduction view switched to ExamNavigationSidebarComponent — so the
location was a leftover from when this bar *was* the student navigation.

Moved to manage/student-exams/student-exam-timeline/exam-navigation-bar/, beside
the consumer and alongside the existing programming-exam-diff child folder.
Pure relocation: no behaviour change.

The component's two relative imports (../events/button/exam-live-events-button
and ../timer/exam-timer) pointed into exam/overview/ and would have broken, so
they are now absolute app/exam/overview/... paths. templateUrl and styleUrls are
relative to the component and moved with it. The `$exam-navigation-bar-*` theme
variables are global SCSS names, not paths, so they are unaffected.

Verified: student-exam-timeline and the bar's own specs pass (3 files / 80
tests), the whole exam module passes (84 files / 1017 tests), `pnpm run compile`
and `compile:tests` both exit 0 — either would fail on a missed import path —
and eslint reports 0 errors on app/exam.
