+++ src/main/webapp/app/exam/manage/student-exams/student-exam-timeline/exam-navigation-bar/exam-navigation-bar.component.scss
