+++ src/main/webapp/app/exam/manage/student-exams/student-exam-timeline/exam-navigation-bar/exam-navigation-bar.component.ts
@@ -17,11 +17,11 @@ import { faBars, faCheck, faEdit } from '@fortawesome/free-solid-svg-icons';
 import { ProgrammingSubmission } from 'app/programming/shared/entities/programming-submission.model';
 import { FileUploadSubmission } from 'app/fileupload/shared/entities/file-upload-submission.model';
 import { TranslateDirective } from 'app/foundation/language/translate.directive';
-import { ExamLiveEventsButtonComponent } from '../events/button/exam-live-events-button.component';
+import { ExamLiveEventsButtonComponent } from 'app/exam/overview/events/button/exam-live-events-button.component';
 import { NgClass } from '@angular/common';
 import { NgbTooltip } from '@ng-bootstrap/ng-bootstrap';
 import { FaIconComponent } from '@fortawesome/angular-fontawesome';
-import { ExamTimerComponent } from '../timer/exam-timer.component';
+import { ExamTimerComponent } from 'app/exam/overview/timer/exam-timer.component';
 import { ArtemisTranslatePipe } from 'app/foundation/pipes/artemis-translate.pipe';
 import { SubmissionVersion } from 'app/exam/shared/entities/submission-version.model';
