+++ src/main/webapp/app/exam/manage/student-exams/student-exam-timeline/student-exam-timeline.component.spec.ts
@@ -15,7 +15,7 @@ import { QuizExamSubmissionComponent } from 'app/exam/overview/exercises/quiz/qu
 import { FileUploadExamSubmissionComponent } from 'app/exam/overview/exercises/file-upload/file-upload-exam-submission.component';
 import { ArtemisDatePipe } from 'app/foundation/pipes/artemis-date.pipe';
 import { ArtemisTranslatePipe } from 'app/foundation/pipes/artemis-translate.pipe';
-import { ExamNavigationBarComponent } from 'app/exam/overview/exam-navigation-bar/exam-navigation-bar.component';
+import { ExamNavigationBarComponent } from 'app/exam/manage/student-exams/student-exam-timeline/exam-navigation-bar/exam-navigation-bar.component';
 import { MockTranslateValuesDirective } from 'test/helpers/mocks/directive/mock-translate-values.directive';
 import { EntityArrayResponseType, SubmissionService } from 'app/exercise/submission/submission.service';
 import { TextExercise } from 'app/text/shared/entities/text-exercise.model';
